Moved client credentials auth to oauth endpoint without versioning
